+++ packages/styles/src/packages/atomic/button/rarui-button.css.ts
@@ -30,18 +30,27 @@ const button = recipe({
       color: varsThemeBase.colors.content.disabled,
       cursor: "not-allowed",
     },
+    ":focus-within": {
+      outline: "none",
+      boxShadow: varsThemeBase.utils["focus-ring"],
+    },
     ":after": {
       transition: "opacity 150ms ease",
       content: "",
       position: "absolute",
       opacity: 0,
       inset: -1,
-      backgroundColor: varsThemeBase.colors.surface["on-brand-hover"],
+      backgroundColor: varsThemeBase.colors.surface["brand-hover"],
       borderRadius: varsThemeBase.shape.border.radius["2xs"],
     },
     selectors: {
       "&:hover:after": {
         opacity: 1,
+      },
+      "&:active:after": {
+        opacity: 1,
+        backgroundColor: varsThemeBase.colors.surface["brand-press"],
+        borderColor: varsThemeBase.colors.surface["on-brand-press"],
       },
     },
   },
@@ -120,7 +129,22 @@ const button = recipe({
      * Defines the visual variant of the badge, affecting its background style, border and text.
      */
     variant: {
-      solid: {},
+      solid: {
+        ":after": {
+          transition: "opacity 150ms ease",
+          content: "",
+          position: "absolute",
+          opacity: 0,
+          inset: -1,
+          backgroundColor: varsThemeBase.colors.surface["on-brand-hover"],
+          borderRadius: varsThemeBase.shape.border.radius["2xs"],
+        },
+        selectors: {
+          "&:hover:after": {
+            opacity: 1,
+          },
+        },
+      },
       outlined: {
         backgroundColor: "transparent",
       },