+++ packages/styles/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 RarUI Styles deprive all styles needed to build components.
 
+## 2024-06-12 `1.28.1`
+
+#### 🐛 Bug fixes
+
+- Added fixed height for `input` component on small, medium and large size. ([#53](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/53) by [@vinicius.osmainchi](https://git.rarolabs.com.br/vinicius.osmainchi))
+
+- Fixed the hover color for solid variation on `button` component. ([#53](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/53) by [@vinicius.osmainchi](https://git.rarolabs.com.br/vinicius.osmainchi))
+
+#### 💡 Others
+
+- Added focus and pressed state on `button` component. ([#53](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/53) by [@vinicius.osmainchi](https://git.rarolabs.com.br/vinicius.osmainchi))
+
+- Added transition for hover and pressed state on `iconButton` component. ([#53](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/53) by [@vinicius.osmainchi](https://git.rarolabs.com.br/vinicius.osmainchi))
+
 ## 2024-06-12 `1.28.0`
 
 #### 🎉 New features
@@ -36,6 +50,12 @@ RarUI Styles deprive all styles needed to build components.
 #### 🎉 New features
 
 - Added new style pack for `accordion` component. ([#52](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/52) by [@junior](https://git.rarolabs.com.br/junior))
+
+## 2024-06-10 `1.24.1`
+
+#### 🐛 Bug fixes
+
+- Adjusting height for medium size and padding of the icon container for the `input` component on small, medium and large size. ([#52](https://git.rarolabs.com.br/frontend/rarui/-/merge_requests/52) by [@vinicius.osmainchi](https://git.rarolabs.com.br/vinicius.osmainchi))
 
 ## 2024-06-02 `1.24.0`